coding/.backup_coding_20251020_133110/pyproject.toml
[project]
name = "coding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.13.1",
    "google-adk[eval]>=1.16.0",
    "google-cloud-trace>=1.13.5",
    "openinference-instrumentation-google-adk>=0.1.6",
]
